KP-7131: Renamed numeric to Decimal

+++ src/Core/SharedKernels/Questionnaire/WB.Core.SharedKernels.Questionnaire/Documents/VariableType.cs
@@ -3,7 +3,7 @@
     public enum VariableType
     {
         Integer,
-        Numeric,
+        Long,
         Boolean,
         DateTime,
         String
